+++ threedi_schematisation_editor/vector_data_importer/importers.py
@@ -6,7 +6,8 @@ from qgis.core import QgsCoordinateTransform, QgsPointLocator, QgsProject
 from threedi_schematisation_editor import data_models as dm
 from threedi_schematisation_editor.utils import gpkg_layer
 from threedi_schematisation_editor.vector_data_importer.integrators import (
-    LinearIntegrator,
+    ChannelIntegrator,
+    PipeIntegrator,
 )
 from threedi_schematisation_editor.vector_data_importer.processors import (
     ConnectionNodeProcessor,
@@ -147,7 +148,7 @@ class LinesImporter(Importer):
         target_model_cls,
         target_layer=None,
         node_layer=None,
-        channel_layer=None,
+        conduit_layer=None,
         cross_section_location_layer=None,
     ):
         super().__init__(
@@ -163,10 +164,15 @@ class LinesImporter(Importer):
             self.fields_configurations,
             self.conversion_settings,
         )
-        if self.conversion_settings.integrate:
-            self.integrator = LinearIntegrator.from_importer(
-                dm.Channel, channel_layer, cross_section_location_layer, self
+        if self.conversion_settings.integrate_channels:
+            self.integrator = ChannelIntegrator.from_importer(
+                conduit_layer, cross_section_location_layer, self
             )
+        elif self.conversion_settings.integrate_pipes and self.target_model_cls in [
+            dm.Weir,
+            dm.Orifice,
+        ]:
+            self.integrator = PipeIntegrator.from_importer(conduit_layer, None, self)
 
 
 class CulvertsImporter(LinesImporter):
@@ -177,7 +183,7 @@ class CulvertsImporter(LinesImporter):
         *args,
         structure_layer=None,
         node_layer=None,
-        channel_layer=None,
+        conduit_layer=None,
         cross_section_location_layer=None,
     ):
         super().__init__(
@@ -185,7 +191,7 @@ class CulvertsImporter(LinesImporter):
             target_model_cls=dm.Culvert,
             target_layer=structure_layer,
             node_layer=node_layer,
-            channel_layer=channel_layer,
+            conduit_layer=conduit_layer,
             cross_section_location_layer=cross_section_location_layer,
         )
 
@@ -198,7 +204,7 @@ class OrificesImporter(LinesImporter):
         *args,
         structure_layer=None,
         node_layer=None,
-        channel_layer=None,
+        conduit_layer=None,
         cross_section_location_layer=None,
     ):
         super().__init__(
@@ -206,7 +212,7 @@ class OrificesImporter(LinesImporter):
             target_model_cls=dm.Orifice,
             target_layer=structure_layer,
             node_layer=node_layer,
-            cchannel_layer=channel_layer,
+            conduit_layer=conduit_layer,
             cross_section_location_layer=cross_section_location_layer,
         )
 
@@ -219,7 +225,7 @@ class WeirsImporter(LinesImporter):
         *args,
         structure_layer=None,
         node_layer=None,
-        channel_layer=None,
+        conduit_layer=None,
         cross_section_location_layer=None,
     ):
         super().__init__(
@@ -227,7 +233,7 @@ class WeirsImporter(LinesImporter):
             target_model_cls=dm.Weir,
             target_layer=structure_layer,
             node_layer=node_layer,
-            channel_layer=channel_layer,
+            conduit_layer=conduit_layer,
             cross_section_location_layer=cross_section_location_layer,
         )