+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 - Add importer for cross section locations (#252)
 - Fix adding nodes of adjacent channels (#428)
 - Only create connection nodes when `create_connection_nodes` is set to True
+- Add option to integrate weirs and orifices on pipes (#246)
 
 
 2.3.1 (2025-07-16)